+++ src/GivingCircle.Api.DataAccess/PostgresClientConfiguration.cs
