+++ src/GivingCircle.Api.DataAccess/Properties/AssemblyInfo.cs
