+++ src/GivingCircle.Api.DataAccess/Repositories/IFundraiserRepository.cs
@@ -1,6 +1,11 @@
-﻿namespace GivingCircle.Api.DataAccess.Repositories
+﻿using GivingCircle.Api.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace GivingCircle.Api.DataAccess.Repositories
 {
-    public class IFundraiserRepository
+    public interface IFundraiserRepository
     {
+        Task<IEnumerable<Fundraiser>> ListAllFundraisersAsync();
     }
 }
