--- src/GivingCircle.Api.DataAccessLibrary/Repositories/IFundraiserRepository.cs
@@ -1,11 +1,0 @@
-﻿using GivingCircle.Api.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace GivingCircle.Api.DataAccess.Repositories
-{
-    public interface IFundraiserRepository
-    {
-        Task<IEnumerable<Fundraiser>> ListAllFundraisersAsync();
-    }
-}
