+++ src/GivingCircle.Api/GivingCircle.Api.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\GivingCircle.Api.DataAccessLibrary\GivingCircle.Api.DataAccess.csproj" />
+    <ProjectReference Include="..\GivingCircle.Api.DataAccess\GivingCircle.Api.DataAccess.csproj" />
     <ProjectReference Include="..\GivingCircle.Api.Models\GivingCircle.Api.Models.csproj" />
   </ItemGroup>
